Fix compile warning in gdk_cursor_new_from_pixmap
(cherry picked from commit f3a5bf48d4e3131ce6087afad5e25bdfae1c5084)

+++ gdk/quartz/gdkcursor-quartz.c
@@ -249,7 +249,7 @@ gdk_cursor_new_from_pixmap (GdkPixmap      *source,
   NSImage *image;
   NSCursor *nscursor;
   GdkCursor *cursor;
-  NSInteger width, height;
+  int width, height;
   gint tmp_x, tmp_y;
   guchar *dst_data, *mask_data, *src_data;
   guchar *mask_start, *src_start;
@@ -265,7 +265,7 @@ gdk_cursor_new_from_pixmap (GdkPixmap      *source,
   gdk_drawable_get_size (source, &width, &height);
 
   bitmap_rep = [[NSBitmapImageRep alloc] initWithBitmapDataPlanes:NULL
-		pixelsWide:width pixelsHigh:height
+		pixelsWide:(NSInteger)width pixelsHigh:(NSInteger)height
 		bitsPerSample:8 samplesPerPixel:4
 		hasAlpha:YES isPlanar:NO colorSpaceName:NSDeviceRGBColorSpace
 		bytesPerRow:0 bitsPerPixel:0];
